Make "<Origin>" elements stricter
These were defined as being available to all Animation subtypes, however
only widget script APIs exist for controlling origins on Scale and
Rotation animations.

It's _probably_ not correct as-such to have "<Origin>" be present on
all animations, so it's been removed from the base "<Animation>" element
and only made available to "<Rotation>" and "<Scale>".